Move `ClientSideDataRow` and `HostStatusRow` to own files
