+++ XenAdmin/XenAdmin.csproj
@@ -794,6 +794,8 @@
     <Compile Include="Dialogs\InstallCertificateDialog.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Include="Wizards\BugToolWizardFiles\StatusReportRows\ClientSideDataRow.cs" />
+    <Compile Include="Wizards\BugToolWizardFiles\StatusReportRows\HostStatusRow.cs" />
     <Compile Include="Wizards\BugToolWizardFiles\StatusReportRows\StatusReportRow.cs" />
     <Compile Include="Wizards\ConversionWizard\ConversionWizard.cs">
       <SubType>Form</SubType>
